+++ src/index.js
@@ -3,5 +3,6 @@ import NestedForm from './nested-form'
 import CheckboxSelectAll from './checkbox-select-all'
 import Remote from './remote'
 import ContentLoader from './content-loader'
+import TextareaAutogrow from './textarea-autogrow'
 
-export { Reveal, NestedForm, CheckboxSelectAll, Remote, ContentLoader }
+export { Reveal, NestedForm, CheckboxSelectAll, Remote, ContentLoader, TextareaAutogrow }