+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Adding [content-loader controller](https://github.com/guillaumebriday/stimulus-components/tree/master/src/content-loader)
+- Adding [textarea-autogrow controller](https://github.com/guillaumebriday/stimulus-components/tree/master/src/textarea-autogrow)
 
 ### Change